+++ ClientRequest.hpp
@@ -6,7 +6,7 @@
 /*   By: orezek <orezek@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/16 21:25:45 by orezek            #+#    #+#             */
-/*   Updated: 2024/09/17 22:55:36 by orezek           ###   ########.fr       */
+/*   Updated: 2024/09/17 23:14:53 by orezek           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ class ClientRequest
 {
 	public:
 		ClientRequest();
-		ClientRequest(int &clientFd, int &bytesReceived, char *data);
+		ClientRequest(int &clientFd, ssize_t &bytesReceived, char *data);
 		ClientRequest(const ClientRequest &obj);
 		ClientRequest &operator=(const ClientRequest &obj);
 		~ClientRequest();
